+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Nuvola SDK Change Log
 
 * new-project, convert-project: Fix zero modification times for files copied from a template inside Nuvola ADK.
   Issue: tiliado/nuvolasdk#10
+* check-project: Files with zero modification times are detected. Issue: tiliado/nuvolasdk#10
 
 4.14.0 - November 11th, 2018
 ----------------------------